Aerosmith Rate Table 1270vs/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ./requirements.in
#
pyserial==3.5
    # via -r ./requirements.in
